+++ AdminPortal/Services/SQLite.cs
@@ -1,17 +1,10 @@
 ﻿using Admin.Portal.API.Core.Const;
 using Admin.Portal.API.Core.Enum;
 using Admin.Portal.API.Core.Models;
-using Admin.Portal.API.Core.Models.Base;
 using Admin.Portal.API.Core.Request;
-using Admin.Portal.API.Core.Response;
-using Admin.Portal.API.Filters;
 using Admin.Portal.API.Helpers;
 using Admin.Portal.API.Interfaces;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
-using System;
-using System.Linq;
-using System.Security.Claims;
 
 namespace Admin.Portal.API.Services
 {
@@ -62,6 +55,9 @@ namespace Admin.Portal.API.Services
 
         public async Task<(bool, TenantModel)> CreateTenant(TenantModel context)
         {
+            if (dbContext.Tenants.Where(t => t.Name.ToLower() == context.Name.ToLower()).ToList().Count > 1)
+                throw new Exception(Messages.ERROR_TENANT_EXSIST);
+
             dbContext.Tenants.Add(context);
             await dbContext.SaveChangesAsync(); 
             return (true, context);
@@ -69,6 +65,9 @@ namespace Admin.Portal.API.Services
 
         public async Task<(bool, UserModel)> CreateUser(UserModel context)
         {
+            if (dbContext.Users.Where(u => u.Email.ToLower() == context.Email.ToLower()).ToList().Count > 1)
+                return (false, null);
+
             dbContext.Users.Add(context);
             await dbContext.SaveChangesAsync();
             return (true, dbContext.Users.Where(u => u.Email == context.Email).FirstOrDefault());
@@ -76,6 +75,9 @@ namespace Admin.Portal.API.Services
 
         public async Task<(bool, UserModel)> UpdateUser(UserModel context)
         {
+            if (dbContext.Users.Where(u => u.ID == context.ID).ToList().Count > 1)
+                return (false, null);
+
             dbContext.Users.Update(context);
             await dbContext.SaveChangesAsync();
             return (true, context);
@@ -86,6 +88,9 @@ namespace Admin.Portal.API.Services
             UserModel user = dbContext.Users.Where(u => u.ID == context.UserID).FirstOrDefault();
             user.Tenants.Add(context.TenantID);
 
+            if (user.Tenants.Where(u => (int)u == context.TenantID).ToList().Count > 1)
+                return (false, null);
+
             dbContext.Users.Update(user);
             await dbContext.SaveChangesAsync();
             return (true, user);
@@ -94,6 +99,10 @@ namespace Admin.Portal.API.Services
         public async Task<(bool, UserModel)> UnLinkTenantFromUser(UserTenantLinkRequest context)
         {
             UserModel user = dbContext.Users.Where(u => u.ID == context.UserID).FirstOrDefault();
+
+            if (user.Tenants.Where(u => (int)u == context.TenantID).ToList().Count == 0)
+                return (false, null);
+
             user.Tenants.Remove(context.TenantID);
             dbContext.Users.Update(user);
             await dbContext.SaveChangesAsync();
@@ -102,6 +111,9 @@ namespace Admin.Portal.API.Services
 
         public async Task<bool> DeleteUser(int id)
         {
+            if (dbContext.Users.Where(u => u.ID == id).ToList().Count == 0)
+                return (false);
+
             dbContext.Users.Where(u => u.ID == id).ExecuteDelete();
             await dbContext.SaveChangesAsync();
             return true;
@@ -109,6 +121,9 @@ namespace Admin.Portal.API.Services
 
         public async Task<(bool, TenantModel)> UpdateTenant(TenantModel context)
         {
+            if (dbContext.Users.Where(u => u.ID == context.ID).ToList().Count == 0)
+                return (false, null);
+
             dbContext.Tenants.Update(context);
             await dbContext.SaveChangesAsync();
             return (true, context);
@@ -116,6 +131,14 @@ namespace Admin.Portal.API.Services
 
         public async Task<bool> DeleteTenant(int id)
         {
+            // User Check
+            if (dbContext.Users.Where(u => u.Tenants.Contains(id)).ToList().Count > 0)
+                return (false);
+
+            //Role Check
+            if (dbContext.Roles.Where(u => u.TenantID == id).ToList().Count > 0)
+                return (false);
+
             dbContext.Tenants.Where(t => t.ID == id).ExecuteDelete();
             await dbContext.SaveChangesAsync();
             return true;
@@ -123,17 +146,17 @@ namespace Admin.Portal.API.Services
 
         public async Task<(bool, List<RoleModel>)> GetRoles(int? tenantID)
         {
-            if (tenantID != null)
-            {
-                List<int> roleList = dbContext.Tenants.Where(u => u.ID == tenantID).FirstOrDefault().Roles;
-                return (true, dbContext.Roles.Where(t => roleList.Contains(t.Id)).ToListAsync().Result);
-            }
+            if (tenantID != null)            
+                return (true, dbContext.Roles.Where(u => u.TenantID == tenantID).ToList());
             else
                 return (true, await dbContext.Roles.ToListAsync());
         }
 
         public async Task<(bool, RoleModel)> CreateRole(RoleModel context)
         {
+            if (dbContext.Roles.Where(u => u.Name == context.Name && u.TenantID == context.TenantID).ToList().Count > 0)
+                return (false, null);
+
             dbContext.Roles.Add(context);
             await dbContext.SaveChangesAsync();
             return (true, context);
@@ -141,6 +164,9 @@ namespace Admin.Portal.API.Services
 
         public async Task<(bool, RoleModel)> UpdateRole(RoleModel context)
         {
+            if (dbContext.Roles.Where(u => u.ID == context.ID).ToList().Count == 0)
+                return (false, null);
+
             dbContext.Roles.Update(context);
             await dbContext.SaveChangesAsync();
             return (true, context);
@@ -148,18 +174,22 @@ namespace Admin.Portal.API.Services
 
         public async Task<bool> DeleteRole(int? Id)
         {
-            dbContext.Roles.Where(t=> t.Id == Id).ExecuteDelete();
+            if (dbContext.Roles.Where(u => u.ID == Id).ToList().Count == 0)
+                return (false);
+
+            dbContext.Roles.Where(t=> t.ID == Id).ExecuteDelete();
             await dbContext.SaveChangesAsync();
             return (true);
         }
 
         public async Task<(bool, RoleModel)> LinkClaimToRole(RoleClaimLinkRequest context)
         {
-            RoleModel role = dbContext.Roles.Where(u => u.Id == context.RoleId).FirstOrDefault();
+            RoleModel role = dbContext.Roles.Where(u => u.ID == context.RoleId).FirstOrDefault();
             if (role.Claims == null) role.Claims = new();
             foreach (int claim in context.ClaimIds )
             {
-                role.Claims.Add(claim);
+                if(!role.Claims.Contains(claim))
+                    role.Claims.Add(claim);
             }
             
             dbContext.Roles.Update(role);
@@ -169,60 +199,54 @@ namespace Admin.Portal.API.Services
 
         public async Task<(bool, RoleModel)> UnLinkClaimFromRole(RoleClaimLinkRequest context)
         {
-            RoleModel role = dbContext.Roles.Where(t => t.Id == context.RoleId).FirstOrDefault();
+            RoleModel role = dbContext.Roles.Where(t => t.ID == context.RoleId).FirstOrDefault();
+
             foreach (int claim in context.ClaimIds) 
             {
-                role.Claims.Remove(claim);
+                if (role.Claims.Contains(claim))
+                    role.Claims.Remove(claim);
             }
             dbContext.Roles.Update(role);
             await dbContext.SaveChangesAsync();
             return (true, role);
         }
 
-        public async Task<(bool, TenantModel)> LinkRoleToTenant(TenantRoleRequest context)
-        {
-            TenantModel tenant = dbContext.Tenants.Where(t=>t.ID == context.TenantId).FirstOrDefault();
-            if (tenant.Roles == null) tenant.Roles = new();
-            foreach(int role in context.RoleIds)
-            {
-                tenant.Roles.Add(role);
-            }
-            dbContext.Tenants.Update(tenant);
-            await dbContext.SaveChangesAsync();
-            return (true, tenant);
-        }
+        //public async Task<(bool, TenantModel)> LinkRoleToTenant(TenantRoleRequest context)
+        //{
+        //    TenantModel tenant = dbContext.Tenants.Where(t=>t.ID == context.TenantId).FirstOrDefault();
+        //    if (tenant.Roles == null) tenant.Roles = new();
+        //    foreach(int role in context.RoleIds)
+        //    {
+        //        tenant.Roles.Add(role);
+        //    }
+        //    dbContext.Tenants.Update(tenant);
+        //    await dbContext.SaveChangesAsync();
+        //    return (true, tenant);
+        //}
 
-        public async Task<(bool, TenantModel)> UnLinkRoleFromTenant(TenantRoleRequest context)
-        {
-            TenantModel tenant = dbContext.Tenants.Where(t => t.ID == context.TenantId).FirstOrDefault();
-            foreach (int role in context.RoleIds)
-            {
-                tenant.Roles.Remove(role);
-            }
-            dbContext.Tenants.Remove(tenant);
-            await dbContext.SaveChangesAsync();
-            return (true, tenant);
-        }
+        //public async Task<(bool, TenantModel)> UnLinkRoleFromTenant(TenantRoleRequest context)
+        //{
+        //    TenantModel tenant = dbContext.Tenants.Where(t => t.ID == context.TenantId).FirstOrDefault();
+        //    foreach (int role in context.RoleIds)
+        //    {
+        //        tenant.Roles.Remove(role);
+        //    }
+        //    dbContext.Tenants.Remove(tenant);
+        //    await dbContext.SaveChangesAsync();
+        //    return (true, tenant);
+        //}
 
-        public async Task<List<string>> GetUserClaims(int userID)
+        public async Task<List<RoleClaimModel>> GetUserClaims(int userID)
         {
             List<int> tenants = dbContext.Users.Where(u => u.ID == userID).FirstOrDefault().Tenants;
-            List<List<int>> roles = dbContext.Tenants.Where(t => tenants.Contains(t.ID)).Select(t => t.Roles).ToList();
-            List<int> uniqueRoles = [];
-            foreach (List<int> x in roles)
-                foreach (int item in x)
-                    if (!uniqueRoles.Contains(item))
-                        uniqueRoles.Add(item);
+            List<RoleModel> roles = dbContext.Roles.Where(r => tenants.Contains(r.TenantID)).ToList();
 
+            List<RoleClaimModel> claims = [];
+            foreach (RoleModel role in roles)
+                foreach (int claim in role.Claims)
+                    claims.Add(new() { ID = role.ID, TenantID = role.TenantID, Claim = dbContext.Claims.Where(c => c.Id == claim).FirstOrDefault() });
 
-            List<List<int>> roleClaims = dbContext.Roles.Where(r => uniqueRoles.Contains(r.Id)).Select(c => c.Claims).ToList();
-            List<int> uniqueClaims = [];
-            foreach (List<int> x in roleClaims)
-                foreach (int item in x)
-                    if (!uniqueClaims.Contains(item))
-                        uniqueClaims.Add(item);
-            
-            return dbContext.Claims.Where(c => uniqueClaims.Contains(c.Id)).Select(c => c.Name).ToList();
+            return claims;
         }
         public async Task<List<ClaimModel>> GetClaims()
         {
